feat: allow specifying auto_start as a CLO
This is specifically for the use-case of wanting to have auto_start
enabled without being able to edit qw.rc due to it being a submodule.

This also adds tile_skip_title = true which is just a small QOL thing
when running qw with tiles. It doesn't affect console or anything else,
it just skips a "Loading complete, press any key to start" message.