feat: Move qw Lua to a separate file
This commit moves qw's lua into `qw.lua` and provides instructions in
`qw.rc` for how to include the lua file when running crawl. The benefit
of this is allowing the editing of qw's complicated Lua in an editor's
Lua mode without requiring a special syntax file.

This commit updates README.md with updated instructions for running qw
and also updates various outdated information.