feat: Don't consider firewood as enemies
qw currently has trouble with movement through water when plants and
other firewood monsters are nearby. This commit has it not consider such
monsters as threats.

This fixes the water movement issue, but may lead to qw getting stuck
when cutting through such monsters is the only way to proceed. Simply
having qw attempt to destroy all firewood as it's encountered is a crude
approach to deal with this issue and may lead to many additional
deathsif it does so when a level isn't fully cleared. An ideal solution
would have qw conditionally re-explore a level in a mode that has it
destroy such monsters, but this would have to be implemented with care,
since there are a number of reasons why a level might be partially
unexplored (runed doors, transporters, dangerous cloud generators, etc).