feat: Update weapon brand evaluations
Update for the combination of the many names of the vorpal brand into
simply "vorpal". Also add a reasonable evaluation of the protection
brand. It's certainly not a top tier brand, but its value isn't 0, so we
go with 50, the same value assigned to venom.

Add an evaluation for the spectral brand that's the same as
electrocution. Completely non-rigorous observational evidence supports
the idea that qw does well with spectral, so we can try this for now.