feat: Equipment consideration updates
Don't equip armour with the harm property, either from artefacts or the
scarf of harm. Amulets of harm no longer exist, so until this commit qw
was happy to wear said items. It's not actually clear that harm is a
net-negative for qw; two of its best runs have used harm items,
including its farthest reaching run so far. However those runs involved
qw having outstanding items in other categories. Given that harm
significantly increases the danger qw faces, particularly from certain
ranged attacks that don't check many or any of its defensive stats, it's
best to disable use of harm items until that can be tested in some
controlled fashion.

Add an evaluation of Regen and Acrobat properties, and tweaks values to
put them on the same level as GSpirit. Regen was considered for amulets
of regeneration only, and Acrobat wasn't considered at all. Add Regen as
a linear property with a value of 100, making one level worth twice that
of one point of Slay/AC/EV, and give the same value to the Acrobat and
GSpirit properties. The regeneration property has been buffed since the
old value of 75 was given, so it seems reasonable to increase its value.
Acrobat is equivalent to +15 when running, something qw likes to do
probably excessively, so it seems reasonable that it should evaluate at
least as well as two points of EV. We make GSpirit have 100 because
there seems to be a consistency to how the qw scoring weighs similarly
functioning properties with the same values.

We also refactor handling of the reflection property, allowing it to be
handled in the same general way, which also allows evaluation of shields
with the reflection ego. Some logic in `resist_vec()` used for for
`resist_dominated()` seems to examine only some properties like
rF/rC/Spirit without properly considering linear properties like AC and
SH. Hence I've added Acrobat to `resist_vec()` but not Regen until I can
better understand how this function is actually used.

Some of the usual reindent and spacing cleanups for changed or nearby
code are also included.