feat: Generalize the plan code
This commit makes an ENDGAME_PLANS table that defines a set of plans
that can be used, giving definitions for "3runes", "15runes", "all"
(15runes+zig), and "zig" (Zig after doing some easier extended
branches). It makes the default plan be "3runes", which is the usual
plan list of "zot". Additionally there's syntax handling for
COMBO_CYCLE_LIST to parse a "!<plan-name>" suffix after the existing
"^<god-initials>" suffix syntax. Hence combos in the cycle list can
specify a plan, and qw will follow this plan when that combo is selected
from the cycle. This plan is stored in `c_persist`, like we do already
for the combo god list, so that the plan will be followed even if qw's
game is saved and reloaded.

This allows us to rotate qw through e.g. both 3 rune and 15 rune
attempts without having to run multiple instances configured with
different plans. We put some examples of plan usage in the default
COMBO_CYCLE_LIST, additionally adding an example of multiple god
initials. Documentation comments for the variables involved are also
improved.