feat: Reorganize and update qw.rc
Move the settings into sections, each with comment headers. Move all
setup instructions to the top in the opening comment section. Change the
default combo to GrBe, since this is presently qw's best combo. Remove
some outdated options. Move the list of qw accomplishments to
accomplishments.md, so we don't have to change the rc file when updating
this.