feat: Consolidate qw debugging output
Make a DEBUG_MODE variable, defaulting to false, that enables various qw
debugging output. This replaces the earlier SPAM variable. Add a `debug`
argument to `say()` that can be set to indicate output should only be
made when debug mode is enabled. Also expand and enable the plan data
debugging statement under this mode. It now prints the plan name as well
as the plan result.