fix: Initial work on stash search improvements
For conversions to TSO, there's existing code to handle attempting to
use the second match in stash searches as a fallback if we fail to
travel to the first result. The current implementation performs an
unwanted movement action if the search returns no results. This commit
begins to implement use of the down arrow key to choose the second
search results instead of using the 'b' key, the latter being a movement
key. We can instead unbind down arrow for movement and use
`crawl.sendkeys()` to send a down arrow key via a keycode. Unfortunately
the current `crawl.sendkeys()` implementation is incorrect, so this code
won't work for now. This code can be revisited once that implementation
is fixed.

Ideally we'd have a full implementation of a clua stash interface so we
wouldn't have to navigate these menus directly. advil has suggested a
simple implementation that only returns the number of results of a
search, which would be sufficient to know when the menu actions would
succeed. Fixing `crawl.sendkeys()` is good to do on its own, but I may
go with his suggestion for a simple stash interface when returning to
this problem.