fix: Some corrections to qw.rc
Use plan names of "3runes" and "15runes" instead of the singular
version. Also add Trog to the default god list. Trog must be included to
use the combo option of GrBe.handaxe, which is the default combo
selection method enabled. Tweak the default combo cycle set to not
specify Trog for berserkers as well.