feat: Updates to monster assessment
qw's monster assessment system is currently very simple, consisting of a
"scary" monster list used to determine when to use to use berserk,
heroism, or buff on the orb run, and a "bia monsters" list that's
actually used for assessment of Brothers in Arms, Finesse, Summon
Greater Servant, Apocalypse, and Summon Divine Warrior.

This commit makes additions to these list based on newer monsters,
especially those added in 0.29-a, as well as data from recent qw runs.
We also rename the `bia_monsters` to `nasty_monsters`, since its use it
not exclusive to Brothers in Arms.