fix: Correct plan handling and refactor
We previously expanded plan handling functionality in e46275e7, but this
introduced a bug where stale plan data from the previous game was used
if it had a non-default plan. Fix this by determining plan data only
after we've handled any first turn initialization at the start of a new
game. Also note the intended plan and god list at the start of the game.
This helps to diagnose plan-related bugs and allows us to know the plans
of an in-progress qw game by looking at the game dump.