feat: Adjustments to use of Ru abilities
Use Apocalypse if we see a scary monster and have no drain, or if we
have drain less than "heavily drained" and we either have 50% or less HP
or see a nasty monster. Previously qw was quite conservative about when
to use Apocalypse. Based on some quick testing, these conditions seem to
result in qw using the ability more while not always being heavily
drained or worse.

Ru's Draw Out Power will be used with priority over Apocalypse if our HP
is 25% or less and we have no worse than light drain. This may not be
the ideal logic in that Draw Out Power might rarely see use. It may be
possible to use this ability in a style closer to Heroism, where it can
be used at e.g. 70% HP if we have no drain. Some future test games
should clear up as to whether this is necessary.

Also make a drain level check function and have all ru ability functions
use this.