feat: Refactoring and cleanups
Generalize the comparison function for the custom `moninfo` frontend
data table returned by its `get_monster_info()` function. We add
functionality to reverse specific fields when making the comparison so
that it can be used in another function. This data table scheme is
awkward and need more looking into.

Also included are some cleanups for spacing and line breaks.