feat: Add version number support
Since qw logic changes over time, having a version string will be useful
to statistical purposes. We need annotated git tags to have proper
version strings, and a script to retrieve this string with `get
describe` and put it somewhere into qw. I usually use the
https://github.com/gammafunk/dcss-rc/blob/master/inline-rc.sh script
from my dcss-rc project to create a final qw.rc for purposes of running
it online. Hence this commit takes this script and specializes it for
purpose of setting up a final qw.rc.

`make-qw-rc.sh` puts the string into a QW_VERSION variable in the input
rc file (default qw.rc) and inlines the input lua file (default qw.lua)
at the end. It has options to change the output file (default qw.rc.out)
and arguments to change the input rc and lua files. qw now notes the
version string on the first turn of the game.