feat: Improve water movement and refactor
Previously it was possible for qw loop trying to move into water and
then immediately back to land when monsters are nearby. This loop can go
on indefinitely when water-bound monsters are nearby that can't reach
us. This commit reworks tactical water stepping to not attempt to move
back onto land unless there's an adjacent threat that would try to stay
adjacent to us as we moved. This way qw will keep moving towards any
water-bound monsters as necessary, only interrupting this to improve its
position when there's a real benefit.

This commit also refactors various radius checks for monsters in the
square assessment process into a single function that passes over the
enemy list once.