feat: Rework qw throwing
qw has basic support for throwing, but this has been broken due to
recent quiver changes. Furthermore this support is only for zero-skill
throwing without much consideration for missile evaluation. This commit
fixes throwing by using the appropriate keys and adds missile evaluation
for purposes of selecting the best one and for consideration of whether
to train throwing skill.

We add throwing skill evaluation that starts at a utility value
corresponding to 1/5th of the value we assign to worth of training a
weapon for min delay. This value doubles or triples if our best missile
is a javelin or a large rock (for Tr/Og), respectively, and is increased
by 50% if our missile has an ego. Hence throwing skill will be trained,
but with less priority than more critical skills like
Weapon/Fighting/Armour/Shields/Dodging.

We still rely on the automatic quiver targeting for ranged attacks and
make no attempt to choose a best target according to qw's threat
evaluation and pathing logic. It would be good to use this logic in some
form, possibly extending it to support wands and spells. However we may
want to look into extending the clua targeter interface so we don't have
to do so much path determination work ourselves. Until then, we have
this commit and can possibly make some relatively simple commits in the
future to add wand/evokable support and some basic target evaluation.