feat: Use and train for Cleansing Flame
This commit teaches qw how to assess whether it wants to use cleansing
flame. The current logic if any of the following are true:

1) We have scary/nasty monsters at range 2 but none at range 1. The idea
   being that it's likely better to cleave if scary/nasty monsters are
   adjacent.

2) There are more than 8 monsters in range 2 with holy vulnerability. At
   this point, even if vulnerable monsters are completely surrounding
   us, we'll hit more monsters with cleansing flame than we will cleave.

3) We are at 50% hp or less, we see hostile monsters, and we'd hit the
   most monsters likely to have TSO restore our HP if we used cleansing
   flame. By this we means there are more non-summoned undead/demons in
   range 2 of us than there are adjacent, and we also require that there be
   at least 4 such monsters. The idea is that we need HP, and cleansing
   flame gives us the best chance to restore some without using
   consumable heal.

We also teach qw to highly value training Invocations while in extended.
It gives it base utility of 1.5, the same as it would training a weapon
to min delay, since cleansing flame the best increase in damage we have
at this point.