feat: Some basic strats for Hells
Have qw buff with might+haste when in one of the Hell branches and we
see 5 or more nasty monsters not vulnerable to any holy wrath we may
have. If the number of such monsters is 9 or more, teleport. This is a
rough first pass at giving qw some better tactics for hells, with more
work to follow. Future work will include blinking to stairs and
generally prioritizing movement to stairs.