doc: Updates to README and accomplishments
Document some current qw win results from 0.29-a and reorganize and
update the list of past accomplishments. Reorganize the README to
suggest offline or private server play and tweak instructions. Simplify
the list of accomplishments to refer to accomplishments.md. Update qw.rc
to call the first set of variables Interface instead of Online, and move
the qw version variable there.