feat: Update some equipment evaluations
Penalize the Obsidian Axe appropriately for cases where we worship Oka
or we have Sacrifice Love under Ru. qw loves this axe in terms of its
utility rating, but it's far less good under Oka now and it  dies
extremely often with it on Vaults:5 in particular. We don't rate it as
never-use in these cases because it remains a fairly good weapon at
least in terms of getting qw to Vaults:5, even if it tends to make qw
spend far too many teleport scrolls. Even when it can make its demon
summons, nerf its +300 rating to +200, since it seems to eventually
cause qw to die in 3-rune portions of the game pretty often.

Also bump the value of the spectralizing brand to 200 over
electrocution's 150, since qw seems to do even better with the former.