feat: Increase value for Will
qw seems to die fairly often from Mark in Vaults and even dies with very
well equipped chars due to paralysis on places like Vaults:5 and Tomb:1.
This suggests we undervalue Will a bit, so let's increase its utility.

Instead of 75 utility per pip of Will up to a max of 200, we give 100
per pip of will up to a max of 300. In Vaults, we give 150 per pip, up
to a max of 450.