feat: Tweak utility for Shields and Throwing
qw seems to be training Throwing a bit high relative to the value it
gets and shield a bit too low. By extended, in a tower shield, it's only
at about 20 Shields, giving it a weapon delay penalty of something like
1.5. It's at about 18 Throwing, so could stand to lose a few levels of
Throwing for Shields and focus Shield more in general. Decrease the
Throwing utility by 33% and increase the delay scaling factor for Shield
utility by 66%.