feat: Improve debugging output
Add debug output channels for `dsay()` and a `DEBUG_CHANNELS` rcfile
variable that can select which channels are enabled. Use these channels
for related debug messages; we currently have channels "main" (the
default), "plans", and "skills". The "plans" channel gives the execution
and results of every plan, and "skills" shows the selected training
skill and its utility. The channel groups and types should be greatly
expanded, but this is a start.