feat: More complete stair exploration
qw can easily get stuck when it travels down stairs to disconnected
areas. Even with the travel code rewrite, it doesn't have a concept of
visiting new stairs to reach previously unreachable areas. This also
leads it to ignore disconnected areas, especially in Orc or branches
like Dungeon, Vaults, and Depths that often have naturally occurring
"cells" due to vault placement or just as a result of the layout.

This commit adds stair tracking via c_persist and support for navigating
to and exploring stairs from levels adjacent to our plan's target
levels.  For a plan to be complete, we now require that we can reach all
stone stairs the level can generate. To travel to staircases, we use map
mode and cycle via the '>' and '<' keys, keeping track of landing
positions and key-presses used so we can guarantee that we're reaching
new staircases. The system is robust in terms of resuming after travel
interruptions and restarting if a save/reload is made.

Ideally we could simplify and improve things by making information about
unexplored stairs available via stash search or at least via the `view`
library. Future work can implement fall-back conditions for completion
should we fail to be able to reach stairs, which might happen in
situations involving stairs on an island of lava or deep water, for
example. It would be good to unify the handling of Tomb to the extent
possible with our new stair searching code. Finally, Slime will need
support for specifying a minimum number of required stairs, so we can
avoid full-exploring that branch.