fix: Perform flight checks for weapon swaps
Don't try to swap weapons if we're over deep water or lava and our
weapon is our sole source of flight. We allow swapping over water if
amphibious or if we would still have flight after the swap. This commit
also refactors flight and deep water/lava terrain handling. We move
evaluation of the flight property into the main item property checking
code, and add better check functions for terrain and intrinsic flight.