feat: Planning code enhancements and refactoring
We'd like to be able to check future plans to visit a given branch
without special tracking variables that have to be updated independently
as the game progresses. This commit rewrites some portions of planning
clode and adds helper functions to that end. We can now check if future
plans we've not yet completed will have us visit a branch, which is
relevant to item property evaluation.

Additionally there are better helper functions to let us refer branch
ends and checking runes without hard-coding depth numbers and rune names
at many locations.