feat: Improved plans, travel, and stair usage
The changes in f8c0728 were a major step, but qw still has major
deficiencies in the logic that governs where it should go. I've
overhauled qw's game plan code (again) to have more consistent logic in
terms of what a game plan represents, how the game plan status is
derived, and to have better separation of concerns between game planning
and travel logic. To help avoid confusion, I refer to the units of
planning that qw uses to reach its goals as game plans to distinguish
them from the plan functions in its plan cascades, since the latter
represent individual actions or moves as opposed to goals; function and
variable names are updated accordingly. This commit is large and covers
many areas described below. Breaking it into smaller commits would be
nice, but a lot of work due to how related the changes are.

We now have a travel search algorithm that predicts the furthest level
qw must go to and begin exploring in order to eventually reach its game
plan level. This allows us immediately travel from where we are to our
destination, even if the current level hasn't been autoexplored. If qw's
travel destination is outside the desired level of its game plan, it
will search only as much as necessary in order to reveal stairs in the
right direction and then take them. Additionally qw can systematically
take unexplored stairs on previous levels when it has become stuck at
its travel destination after autoexploration. This uses the same logic
we use for finding all stairs in order to fully explore game plan
levels.  Game plan and travel are only recalculated when actually
necessary, that is when level location has changed or when a message or
exploration event adds information that might have qw alter its plans.

The game plan code now has support for specifying gods for conversion
(e.g. God:Okawaru) and runes (e.g. Rune:Swamp) as goals. For rune goals,
qw dives to the relevant level, retrieves the rune when found via
autoexploration, and immediately proceeding to the next level. For
ziggurats, the max zig depth can be specified as part of the Zig game
plan, e.g.  Zig:13, with Zig by itself meaning the entire ziggurat.

Portal branches handling has been rewritten so that we track portal
messages on level entry and portal expiration message, updating plans
accordingly. Portal search is integrated with the travel code, so qw can
explore a level systematically until it finds the entry, trying new
stairs as necessary. Portal data has been merged into branch data, with
a separate table to hold portal-specific info.

Debugging has been improved with the addition of functions
toggle_debug_mode() and toggle_debug_channel() to turn debug itself and
specific debug channels on/of, and reset_stairs() to reset qw's stair
knowledge of a given level and/or stair type. Various relatively
expensive debugging statements that create strings are wrapped in `if
DEBUG_MODE ...` clauses so their presence won't impact performance in
non-debug-mode.

We now have a god data table containing various bits of god info that
was previously spread out, along with helper functions to access this
data. I've added an enum function to create enum tables that are used
for constants that previous had to be stored in individual variables.
This commit also includes the usual batch of comment cleanups and
linebreaks.