fix: Generate Lair branch ranges correctly
The previous commit naively tried to calculate Lair branch ranges at
initialization, but the branches we'll actually have don't usually exist
at this time, which is most often the start of the game. Refactor the
Normal plan code to initialize the cached ranges only after we're sure
we've explored all of Lair.