refactor: Upper-case enum names
The lua table enums leave a lot to be desired, since they don't
protection you from typo names evaluating to nil. Until a better
solution is available, help avoid bad lookups by using all-caps for the
name of the enum table in addition to the all-caps we already use for
the values.