feat: Basic transporter navigation
This commit implements taking transporters in Temple and portal
branches. We use a simplified system similar to the map key
search/waypoint approach, but without the waypoints, since those don't
work in all the places we might like to use transporters. Instead we
track instance of taking a transporter and increment a "zone" count. We
attempt to orient relative to the exit after each transporter, so zone
cycles can be detected. We introduce autoexplore tracking of rune doors
and transporters and use this to trigger the transporter plans only when
appropriate. When levels with transporters no longer report those in the
autoexplore result message, we proceed to the exit normally.

The transporter searching by zone is not accurate if a portal has
multiple exits, but map key searching will generally push qw to try all
transporters until they're all explored. Testing with the Golubria
wizlab has shown this to be the case, for example. That said, this
system will need more work before qw will be able to navigate required
transporters in Pan.

We also enable opening of runed doors in portals, since we'd generally
want to do this regardless of transporters.