fix: correct logic for the unexplored stair plan
`plan_go_to_unexplored_stairs()` wasn't verifying that it actually had
work to do in the case it had gameplan stairs searching work beyond the
final level it could travel to, leading it to map travel between stairs
forever. The conditions for this to happen are rare and involve
disconnected areas on the final level and managing to explore all stairs
in the relevant direction leading to the final level before
autoexploring sufficiently. Add a check for remaining unexplored stairs
to this plan so that it can return false, allowing the secondary
autoexplore and backtracking plans to kick in.