feat: Add portal exploration to the rcfile
Define an `ALLOWED_PORTALS` list in the rcfile that contains all the
portals qw is allowed to attempt. Refactor code to check this, removing
the "enabled" flag from the portal data table. We maintain and update
branch data for all portals as they're find, but only choose a level
portal as a gameplan if it's in the allow list. The default list
includes the usual Bazaar, Sewer, and Ossuary, but also adds the new
portals we're testing: Gauntlet, WizLab, and Desolation.