feat: Improve portal tracking
The logic for tracking portals wasn't robust enough to handle all the
cases that can arise, specifically with bazaars. Since bazaars can be
both timed and permanent, potentially with both types on the same level,
we need to accurately detect permanent bazaars and prioritize timed
portals over them when choosing a portal as a gameplan. With this
commit, for each portal type on a level, we use a list and put any
permanent portal first in the list, with a special high turn value.
Search for bazaar portals in a way that won't match permanent ones if
we're looking for the timed portal.