fix: corrections for portal logic
We needed more accurate message matching for a couple portals with short
branch names sufficiently different from the full name. Gauntlets have a
slightly different appearance message than the other portals, so add a
condition for this. Also don't try to expire portal parent data while
we're in them, since we do this ourselves right before we take the exit,
and it might be good to have access to the parent data.