fix: Unbreak god worship
In addition to using the wrong string for "No God" and having broken the
abandonment plan, the logic for wanting and altar and abandonment wasn't
very functional. This commit fixes the logic, removing the former
c_persist variable and normalizing the case when the god list specifies
"No God". This simplifies the check for whether we need an altar, since
we'll always want one if we have no current god unless we're a demigod
or the god list has specified "No God".

The abandonment plan is fixed and its check is likewise simplified. We
add "No God" as a proper entry in the god table, which allows it to be
specified in a god list either by that full name or by the abbreviations
"0" and "None".