feat: Don't flee under opportunity attacks
Before accepting a step as "fleeing", ensure that we'd not get AOO
either from our current position or from the position we'd move to, but
checking the number of adjacent monsters. qw was trying to flee to
stairs over long distances with sometimes dangerous monsters when it
would be far better to stand and fight. This could use better detection
of whether any adjacent monster is actually threatening, but that will
need to wait for a more robust threat assessment system.