feat: Improve stair search and backtracking
It's possible when searching for stairs for a gameplan level from an
adjacent level that we end up in a region disconnected from the
remaining unexplored stairs on the adjacent level. In this case we need
to take alternate stairs to reach the adjacent level, so we can reach
the unexplored stairs to our gameplan level. This means we need
backtracking for gameplan stairs search in addition to the backtracking
we have already for travel search. This commit adds that support to the
backtracking plan, which is now named
`plan_unexplored_stairs_backtrack()`, since it's no longer specific to
travel searches.

Furthermore, we need to terminate the map search key sequences (`X`
followed by one or more `>` or `<`) when they fail to go a new position.
Previously we let the process continue indefinitely with the idea that
it would always eventually reach a new stair to explore, since we are
only initiating the search in cases where we know unexplored stairs
remain. However in disconnected areas, the `X` command won't cycle to
stairs in other disconnected areas, leading to indefinite searches.
Abort the plan when two consecutive search counts lead to the same
position.