refactor: Simplify stair feature logic
When checking if a feature is an upstairs, simply check for stone up
stairs and for branch exits, instead of doing the this but also having a
bunch of exceptions for branches we don't want to stairdance. Add the
excpetions to the stairdance plan instead. This way the upstairs feature
check function has a much simpler implementation.