fix: More improvements to travel stair exploration
The logic for stair exploration and backtracking still wasn't correct
after a877c2b2 and various fixes afterwards. We weren't systematic
enough in finding alternate levels to try in order to reach our desired
stairs, and the differences in the notions of stair search direction
between travel and gameplan code made handling both forms of stair
searching and backtracking difficult.

This commit has travel set a stair direction only when we send qw to the
depth at which we want to search stairs. Previously travel always set
this direction in the direction it wanted to travel without regard for
whether it could potentially find unexplored stairs at its travel
destination. Have logic for determining a best finalized depth and stair
search direction for travel that considers autoexplore state and
availability of unexplored stairs, possibly setting a final depth on a
level adjacent to the original destination and with a stair search
direction. Now gameplan has no independent notion of a stair search,
instead leaving the gameplan depth as the earliest depth still needing
autoexplore or up/downstairs. The process of finalizing depth for
purposes of finding new stairs is left to travel, since this simplifies
calculations and handling of backtracking. For backtracking, set a state
variable and have the depth finalization leave the depth unmodified, but
with a stair search direction to try to reach the adjacent level from
new stairs.

We also consistently prioritize search the depth above the target depth,
then up stairs on the target depth, then down stairs on the target
depth, then the depth below, since this is the safest exploration route.
Additionally, because we never tried to explore stairs in the relevant
direction on the target depth itself, only those in the (opposite)
relevant direction on the adjacent levels. This caused us to potentially
miss stairs if both the target level and an adjacent level had
disconnected areas. This new exploration pattern should reach all stairs
on every level unless there is a sequence of disconnected areas leading
into one another across a set of at least three sequential levels. This
type of stair configuration would be exceedingly rare, if it ever
occurs, but future work could generalize the search pattern to handle
this.

This commit also refactors functions that check c_persist stair data to
have a `dir` argument instead of having upstairs/downstairs versions of
each function. This makes it a lot easier to write general code that can
handle both directions.