feat: Tweak teleport behaviour under Trog
Under Trog, we were teleporting whenever we were slowed post-berserk and
would like to berserk again. This can happen fairly often in
post-berserk situations that are only somewhat dangerous. Namely when
we're at 50% hp and wouldn't likely dip to the danger zone of 25% even
while slowed, since the remaining monsters are weak. This commit applies
the same 25% HP condition for Trog as is applied to other gods.

Hopefully qw can nearly always get in a telport while slowed, and if
that proves to not be the case, we can try partially re-adding the
post-berserk conditions, but still with a more stringent threshold that
won't result in many wasted teleport scrolls.