feat: Reorganize and tweak monster lists
I've redone many entries in the scary and nasty monster lists based on
the last two months of qw games. The updates add newer monsters and to
make the nasty and scary monster lists more consistent with each other.
Namely most monsters that are the nasty list are now also in the scary
list, unless they blink a lot, which causes problems when using blink.
Some entries are moved to be active at later XLs based on the highest XL
where qw has died to the given monster.

The scary and nasty monster lists had their entries grouped by the XL at
which they apply, but with random ordering, which made examining the
lists difficult. I've alphabetized these lists to make future changes
easier.

This monster list system is far too simplistic and will eventually be
replaced with a more robust monster evaluation scheme, but it's what we
have for now.  The goal of these changes is to make the current system
as effective as possible while also being something that can be easilly
maintained.