feat: Basic Ignis support
This commit adds Ignis support in terms of using Fiery Armour and
converting to a god in the god list after Ignis piety is depleted. Most
of the conversion support is in the normal plan, which allows conversion
to any found god in the god list when piety is gone and we've at least
finished Lair. In the future we'd like to support using Scorch and
support Ignis' Foxfire Swarm and Rising Flame abiltiies, but this should
be enough to get qw started on the TrCA Nemelex Choice.