fix: Better stair exploration logic
When finalizing the travel destination based on the gameplan, the
current logic doesn't consider when qw can already reach all required
stairs for a given direction from the gameplan level. If this condition
is met, we have no need of exploring either the adjacent level in this
direction nor the the stairs in this direction on the gameplan level.
After this commit, qw won't needlessly performing stair exploration in
this case.