fix: Always allow off-level travel
qw had a facility to conditionally accept the off-level travel prompt
where it allowed it by default but disabled it for certain plans that
used stash search. I'm not sure why it needed this, but it was
presumably to prevent it from getting stuck or generally confused in
terms of state should a stash search unexpectedly take it to an adjacent
level. In any case, qw now has more robust planning and travel and needs
to always accept this dialogue for stash search travel to work.