feat: Use the travel system to reach altars
Instead of simply performing a stash search to a known altar, consider
where the altar is and use the travel system to try to reach it. The god
conversion gameplan now sets the branch and depth to a known location of
the altar, allowing travel to take us there with a go command before
doing a stash search. This way we can recover from things like shafts
while under conversion gameplans without being stuck when a stash search
would fail. Travel will take us to the furthest reachable level and have
us autoexplore and search unexplored stairs as necessary to eventually
reach a location where the stash search can succeed.