fix: Handle planning updates within a turn
Certain condtions trigger the need for an update to game planning. For
example, an autoexplore may complete without consuming a turn and mark a
level as autoexplored. This updates the state variable that indicates a
planning update is needed, but since no turn is consumed, qw currently
doesn't have a way to see the update until a turn is consumed somehow.
This leads to it often wasting a teleport scroll as it thinks it's
stuck, only to then update the planning the turn after reading teleport.

This commit moves the gameplan update into a single function which is
called both at the usual place in `turn_update()` but now also in the
plan cascade if a plan returns true yet didn't consume a turn. It still
depends on the state variable that indicates we actually need a planning
update.