fix: Handle unexplore stair travel failure
Rare cases can arise where we can't reach the unexplored stairs we'd
like to take when doing a stair search. In this cases we want to allow
autoexplore to kick in, since our wanting a stair search in the first
place will disable use of autexplore. Add the necessary logic to have
`plan_go_to_unexplored_stairs()` re-enable autoexplore if it fails to
act the first time.

This commit also refactors the plan fail count tracking, using separate
variables for stash travel, go travel, and the map search travel.
Previously these all used the same variable, which would mess up state
if multiple plans using it could be active at the same time. This might
(rarely) happen if we had a Zot:5 vault blocked by teleport/dispersal
traps and had seen the Orb, thus allowing the orb stash search travel
plan to trigger, but had somehow then been teleported outside of the
teleport-blocked area and then retreated upstairs. Stash travel would
fail in this case, and we would need a go command to get back to Zot:5,
so we need these plans to have independent failure tracking. If multiple
stash searches need to be independently active at the same time, we
might need to generalize this even further, but for now these changes
will be sufficient.