fix: Better logic for stash search travel
The move of altar searching to properly use travel introduced some bugs
because we didn't properly consider when we need a stash search. This
commit has the travel search logic to know when we're doing a stash
search travel and not try to finalize the destination based on stair
exploration in this case. This way we'll not try to set off-destination
travel destinations based on stair exploration on the target level,
which isn't relevant to stash search travel. We also more accurately
track all stash search cases, so the first autoexplore can be
appropriately disabled.

Finally, reorder plans that use stash search travel, moving them all to
the second explore plan set. Group related plans together more
consistently.