feat: Flee more aggressively to stairs at low XL
For Delver starts in particular, we want to try to reach upstairs even
when hostile monsters are in view. Current qw logic has it prefer to
fight before taking stairs, but this can be fatal at early XLs. This
commit has qw flee upstairs not only whe HP is below 90%, but also when
we're at XL <=8 and we have autoexplore disabled. The latter only
happens when we'd like to be traveling in some way and we haven't fully
explored the current level. These two conditions are a good indicator
for when e.g. Delver should be avoiding combat and taking stairs
instead.